added info about finish Qt4 support; some fixes for types compliance;

+++ src/create_widgets/domain/common_widgets/cpu_widgets/logical_host_cpu.cpp
@@ -24,6 +24,6 @@ LogicalHostCPU::LogicalHostCPU(
 /* public slots */
 void LogicalHostCPU::changeInfoVisibility(int i)
 {
-    info->setVisible( i>cores );
-    icon->setVisible( i>cores );
+    info->setVisible( (uint)i > cores );
+    icon->setVisible( (uint)i > cores );
 }
